+++ src/main/kotlin/com/emberjs/EmberXmlElementDescriptor.kt
@@ -1,16 +1,10 @@
 package com.emberjs
-import com.dmarcotte.handlebars.psi.impl.HbDataImpl
-import com.dmarcotte.handlebars.psi.impl.HbPathImpl
 import com.emberjs.psi.EmberNamedElement
 import com.emberjs.utils.*
 import com.intellij.codeInsight.documentation.DocumentationManager.ORIGINAL_ELEMENT_KEY
-import com.intellij.lang.javascript.psi.JSElement
-import com.intellij.lang.javascript.psi.jsdoc.JSDocComment
 import com.intellij.psi.*
-import com.intellij.psi.impl.file.PsiDirectoryImpl
 import com.intellij.psi.impl.source.html.dtd.HtmlNSDescriptorImpl
 import com.intellij.psi.impl.source.xml.XmlDescriptorUtil
-import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.xml.XmlAttribute
 import com.intellij.psi.xml.XmlTag
 import com.intellij.xml.XmlAttributeDescriptor
@@ -18,18 +12,7 @@ import com.intellij.xml.XmlElementDescriptor
 import com.intellij.xml.XmlElementsGroup
 import com.intellij.xml.XmlNSDescriptor
 
-class ArgData(
-        var value: String = "",
-        var description: String? = null,
-        var reference: AttrPsiReference? = null) {}
 
-class ComponentReferenceData(
-        public val hasSplattributes: Boolean = false,
-        public val yields: MutableList<EmberXmlElementDescriptor.YieldReference> = mutableListOf(),
-        public val args: MutableList<ArgData> = mutableListOf()
-) {
-
-}
 
 class EmberXmlElementDescriptor(private val tag: XmlTag, private val declaration: PsiElement) : XmlElementDescriptor {
     val project = tag.project
@@ -82,25 +65,6 @@ class EmberXmlElementDescriptor(private val tag: XmlTag, private val declaration
         }
     }
 
-    fun getFileByPath(directory: PsiDirectory?, path: String): PsiFile? {
-        if (directory == null) return null
-        var dir: PsiDirectory = directory
-        val parts = path.split("/").toMutableList()
-        while (parts.isNotEmpty()) {
-            val p = parts.removeAt(0)
-            val d: Any? = dir.findSubdirectory(p) ?: dir.findFile(p)
-            if (d is PsiFile) {
-                return d
-            }
-            if (d is PsiDirectory) {
-                dir = d
-                continue
-            }
-            return null
-        }
-        return null
-    }
-
     /**
      * finds yields and data mustache `@xxx`
      * also check .ts/d.ts files for Component<Args>
@@ -116,80 +80,8 @@ class EmberXmlElementDescriptor(private val tag: XmlTag, private val declaration
         }
         f = EmberUtils.followReferences(target)?.containingFile?.originalFile
         val file = f ?: target.containingFile.originalFile
-        var name = file.name.split(".").first()
-        val dir = file.parent as PsiDirectoryImpl?
-        var template: PsiFile? = null
-        var path = ""
-        var parentModule: PsiDirectory? = null
-        val tplArgs = emptyArray<ArgData>().toMutableList()
-        var tplYields = mutableListOf<YieldReference>()
 
-        if (dir != null) {
-            // co-located
-            if (name == "component") {
-                name = "template"
-            }
-            template = dir.findFile("$name.hbs")
-            parentModule = file.parents.find { it is PsiDirectory && it.virtualFile == file.originalVirtualFile?.parentEmberModule} as PsiDirectory?
-            path = file.parents
-                    .takeWhile { it != parentModule }
-                    .reversed()
-                    .map { (it as PsiFileSystemItem).name }
-                    .joinToString("/")
-
-            val fullPathToHbs = path.replace("app/", "addon/") + "/$name.hbs"
-            template = template
-                    ?: getFileByPath(parentModule, fullPathToHbs)
-                    ?: getFileByPath(parentModule, fullPathToHbs.replace("/components/", "/templates/components/"))
-
-            if (template?.node?.psi != null) {
-                val args = PsiTreeUtil.collectElementsOfType(template.node.psi, HbDataImpl::class.java)
-                for (arg in args) {
-                    val argName = arg.text.split(".").first()
-                    if (tplArgs.find { it.value == argName } == null) {
-                        tplArgs.add(ArgData(argName, "", AttrPsiReference(arg)))
-                    }
-                }
-
-                val yields = PsiTreeUtil.collectElements(template.node.psi, { it is HbPathImpl && it.text == "yield" })
-                for (y in yields) {
-                    tplYields.add(YieldReference(y))
-                }
-            }
-        }
-
-
-        if (name == "template") {
-            name = "component"
-        }
-        val hasSplattributes = template?.text?.contains("...attributes") ?: false
-        val fullPathToTs = path.replace("app/", "addon/").replace("/templates/", "/") + "/$name.ts"
-        val fullPathToDts = path.replace("app/", "addon/").replace("/templates/", "/") + "/$name.d.ts"
-        var containingFile = target.containingFile
-        if (containingFile.name.endsWith(".js")) {
-            containingFile = dir?.findFile(containingFile.name.replace(".js", ".d.ts"))
-        }
-        val tsFile = getFileByPath(parentModule, fullPathToTs) ?: getFileByPath(parentModule, fullPathToDts) ?: containingFile
-        val cls = tsFile?.let {EmberUtils.findDefaultExportClass(tsFile)}
-                ?: containingFile?.let {EmberUtils.findDefaultExportClass(containingFile)}
-                ?: target
-        if (cls is JSElement) {
-            val argsElem = EmberUtils.findComponentArgsType(cls)
-            val signatures = argsElem?.properties ?: emptyList()
-            for (sign in signatures) {
-                val comment = sign.memberSource.singleElement?.children?.find { it is JSDocComment }
-//                val s: TypeScriptSingleTypeImpl? = sign.children.find { it is TypeScriptSingleTypeImpl } as TypeScriptSingleTypeImpl?
-                val attr = sign.toString().split(":").last()
-                val data = tplArgs.find { it.value == attr } ?: ArgData()
-                data.value = attr
-                data.reference = AttrPsiReference(sign.memberSource.singleElement!!)
-                data.description = comment?.text ?: ""
-                if (tplArgs.find { it.value == attr } == null) {
-                    tplArgs.add(data)
-                }
-            }
-        }
-        return ComponentReferenceData(hasSplattributes, tplYields, tplArgs)
+        return EmberUtils.getComponentReferenceData(file)
     }
 
     override fun getAttributesDescriptors(context: XmlTag?): Array<out XmlAttributeDescriptor> {